pyproject.toml
[build-system]
requires = ["setuptools>=78.1.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "cross-ai-core"
version         = "0.3.0"
description     = "Multi-provider AI dispatcher with response caching and error handling."
readme          = "README.md"
license         = {text = "MIT"}
requires-python = ">=3.10"
authors         = [{name = "b202i"}]
keywords        = ["ai", "llm", "anthropic", "openai", "gemini", "xai", "perplexity", "multi-provider"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# requests is always required: used for the Perplexity provider and general HTTP.
# Provider SDKs are optional — install only what you need (see below).
dependencies = [
    "requests>=2.32.4",
]

[project.optional-dependencies]
# Install a single provider:
#   pip install "cross-ai-core[anthropic]"
#   pip install "cross-ai-core[gemini]"
#   pip install "cross-ai-core[openai]"    # also covers xAI (Grok)
anthropic = ["anthropic>=0.84.0"]
gemini    = ["google-genai>=1.65.0"]
openai    = ["openai>=1.70.0"]
xai       = ["openai>=1.70.0"]            # xAI (Grok) uses the OpenAI client SDK

# All providers — used by cross (multi-provider by design):
#   pip install "cross-ai-core[all]"
all = [
    "anthropic>=0.84.0",
    "google-genai>=1.65.0",
    "openai>=1.70.0",
]

dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.urls]
Homepage      = "https://github.com/b202i/cross-ai-core"
Repository    = "https://github.com/b202i/cross-ai-core"
"Bug Tracker" = "https://github.com/b202i/cross-ai-core/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["cross_ai_core*"]

[tool.setuptools.package-data]
cross_ai_core = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
